arguments captor moved to the testsupport package
