+++ internal/rules/provider/filesystem/provider_test.go
@@ -33,7 +33,7 @@ import (
 	config2 "github.com/dadrus/heimdall/internal/rules/config"
 	"github.com/dadrus/heimdall/internal/rules/rule/mocks"
 	"github.com/dadrus/heimdall/internal/x"
-	mock2 "github.com/dadrus/heimdall/internal/x/mock"
+	mock2 "github.com/dadrus/heimdall/internal/x/testsupport/mock"
 )
 
 func TestNewProvider(t *testing.T) {
